Rough sketch of templating language
This is a rough draft of what the templating API and DSL might look
like.

I've stubbed out the macros so that the code compiles(-ish). The focus
is not on a working set of macros right now, but sculpting the overall
look and feel for writing actions (name TBD) and asset definitions.

I've also included a JSON file that illustrates how a
`ContractDefinition` might look, which draws together several assets,
actions and Auth specifications.